pyproject.toml
[project]
name = "boilerplate-chatbot"
version = "0.1.0"
description = "A chat application to chat with philosophers."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-secret-manager>=2.23.0",
    "httpx-oauth>=0.16.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
    "stripe>=11.5.0",
    "vertexai>=1.71.1",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
assets = []
build_command_env = []
commit_message = "chore(release): {version}"
commit_parser = "angular"
logging_use_named_masks = true
major_on_zero = false
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
changelog_file = ""
exclude_commit_patterns = []
mode = "init"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "github-actions <actions@github.com>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.6",
]
